pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-workflows"
version = "0.1.0"
description = "Graph-based multi-agent orchestration platform"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "pydantic>=2.12,<3.0",
    "langgraph>=1.0.6,<2.0",
    "langgraph-prebuilt>=1.0.6,<1.1.0",
    "langchain-anthropic>=0.3.0",
    "openai>=2.0",
    "groq>=1.0",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "structlog>=25.0",
    "rich>=14.0",
    "requests>=2.32",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "sse-starlette>=3.3",
    "anyio>=4.0",
    "psycopg[binary]>=3.2",
    "psycopg_pool>=3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.11",
    "mypy>=1.10",
    "pre-commit>=4.0",
    "httpx-sse>=0.4",
    "pytest-cov>=6.0",
]
observability = ["langfuse>=3.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "E402",  # module docstrings after __future__ imports (legacy pattern)
    "E501",  # line too long (handled by formatter for new code; legacy strings exempt)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "postgres: requires PostgreSQL database (DATABASE_URL env var)",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
